refactor(node): Extract `NodePackageManagerType` to a dedicated file
This gives a better overview. While at it, also move the belonging
`hasNonEmptyFile()`.
